refactor: move shadow-dom-wrapper into html-message-renderer package
- reorganize imports and rename shadow-dom-wrapper test file
