+++ src/commons/mail-message-renderer/html-message-renderer/tests/shadow-dom-wrapper.test.tsx
@@ -13,7 +13,7 @@ import '@testing-library/jest-dom';
 import { useUserSettings } from '@zextras/carbonio-shell-ui';
 import * as darkReader from 'darkreader';
 
-import { ShadowDomWrapper } from 'commons/mail-message-renderer/shadow-dom-wrapper';
+import { ShadowDomWrapper } from '../shadow-dom-wrapper';
 
 jest.mock('darkreader', () => ({
 	...jest.requireActual('darkreader'),
